wip: move readableDistance/Area to textutils.js to prevent circular import
This breaks collectstatic.
